Move buildGradingJobStats to component
